migrated to single Config File that contains both server and client side config. Secrets are loaded for server side only from ENV variables
