+++ server/middleware/agentAuth.js
@@ -1,6 +1,6 @@
 const { Agent } = require('../db');
 const logger = require('../logger');
-const { isDevelopment } = require('../envConfig');
+const { isDevelopment } = require('../../client/src/envConfig');
 
 const withAsync = require('./withAsync');
