+++ server/middleware/jwtAuth.js
@@ -1,6 +1,6 @@
 const jwt = require('express-jwt');
 const jwks = require('jwks-rsa');
-const { auth0, isDevelopment }  = require('../envConfig');
+const { auth0, isDevelopment }  = require('../../client/src/envConfig');
 
 const logger = require('../logger');
