+++ server/services/giftpayClient.js
@@ -1,5 +1,5 @@
 const axios = require('axios');
-const { giftPay } = require('../envConfig');
+const { giftPay } = require('../../client/src/envConfig');
 const logger = require('../logger');
 const uuidv4 = require('uuid/v4');
